DPAV-2067-refactored-grpc-client
