--- src/test/java/uk/gov/dbt/ndtp/federator/client/grpc/GRPCClientTest.java
@@ -1,129 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// © Crown Copyright 2025. This work has been developed by the National Digital Twin Programme
-// and is legally attributed to the Department for Business and Trade (UK) as the governing entity.
-
-/*
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package uk.gov.dbt.ndtp.federator.client.grpc;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
-
-import io.grpc.ChannelCredentials;
-import io.grpc.ManagedChannel;
-import java.util.concurrent.TimeUnit;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.mockito.MockedStatic;
-import uk.gov.dbt.ndtp.federator.common.utils.KafkaUtil;
-import uk.gov.dbt.ndtp.federator.common.utils.TestPropertyUtil;
-import uk.gov.dbt.ndtp.secure.agent.sources.kafka.sinks.KafkaSink;
-
-class GRPCClientTest {
-    private static final String SERVER_NAME = "TEST_SERVER_NAME";
-    private static final String TOPIC_PREFIX = "TEST_TOPIC_PREFIX";
-    private static final String TOPIC_NAME = "TEST_TOPIC_NAME";
-
-    @BeforeAll
-    static void beforeAll() {
-        TestPropertyUtil.clearProperties();
-        TestPropertyUtil.setUpProperties("client.properties");
-    }
-
-    @AfterAll
-    static void afterAll() {
-        TestPropertyUtil.clearProperties();
-    }
-
-    @Test
-    void getRedisPrefix() {
-        String clientName = RandomStringUtils.random(10);
-        String serverName = RandomStringUtils.random(10);
-
-        try (MockedStatic<GRPCClient> grpcClientMockedStatic = mockStatic(GRPCClient.class)) {
-            grpcClientMockedStatic
-                    .when(() -> GRPCClient.generateChannel(anyString(), anyInt()))
-                    .thenReturn(mock(io.grpc.ManagedChannel.class));
-
-            try (GRPCClient client = new GRPCClient(clientName, "key", serverName, "host", 1, false, TOPIC_PREFIX)) {
-                String prefix = client.getRedisPrefix();
-                assertEquals(clientName + "-" + serverName, prefix);
-            }
-        }
-    }
-
-    @Test
-    void getSenderWithTopicPrefix() {
-        try (MockedStatic<KafkaUtil> kafka = mockStatic(KafkaUtil.class)) {
-            kafka.when(() -> KafkaUtil.getKafkaSink(any())).thenReturn(mock(KafkaSink.class));
-            GRPCClient.getSender(TOPIC_NAME, TOPIC_PREFIX, SERVER_NAME);
-            kafka.verify(() -> KafkaUtil.getKafkaSink("TEST_TOPIC_PREFIX-TEST_SERVER_NAME-TEST_TOPIC_NAME"));
-        }
-    }
-
-    @Test
-    void getSenderWithoutTopicPrefix() {
-        try (MockedStatic<KafkaUtil> kafka = mockStatic(KafkaUtil.class)) {
-            kafka.when(() -> KafkaUtil.getKafkaSink(any())).thenReturn(mock(KafkaSink.class));
-            GRPCClient.getSender(TOPIC_NAME, "", SERVER_NAME);
-            kafka.verify(() -> KafkaUtil.getKafkaSink("TEST_SERVER_NAME-TEST_TOPIC_NAME"));
-        }
-    }
-
-    @Test
-    void testGenerateSecureChannelIsMocked() {
-        ManagedChannel mockChannel = mock(ManagedChannel.class);
-
-        try (MockedStatic<GRPCClient> grpcClientMockedStatic = mockStatic(GRPCClient.class)) {
-            grpcClientMockedStatic
-                    .when(() -> GRPCClient.generateSecureChannel(anyString(), anyInt(), any(ChannelCredentials.class)))
-                    .thenReturn(mockChannel);
-
-            // Use matchers for all arguments in verify as well
-            ManagedChannel result = GRPCClient.generateSecureChannel("anyHost", 9999, mock(ChannelCredentials.class));
-            assertSame(mockChannel, result);
-
-            grpcClientMockedStatic.verify(
-                    () -> GRPCClient.generateSecureChannel(anyString(), anyInt(), any(ChannelCredentials.class)));
-        }
-    }
-
-    @Test
-    void constructor_nonTls_uses_generateChannel_and_close_shuts_down_channel() throws Exception {
-        ManagedChannel channel = mock(ManagedChannel.class);
-        when(channel.shutdown()).thenReturn(channel);
-        try (MockedStatic<GRPCClient> grpcStatic = mockStatic(GRPCClient.class)) {
-            // Let unspecified static methods call real methods
-            grpcStatic
-                    .when(() -> GRPCClient.generateChannel(anyString(), anyInt()))
-                    .thenReturn(channel);
-            // Build client with TLS disabled so it uses generateChannel
-            GRPCClient client = new GRPCClient("client", "key", "server", "host", 1234, false, "pref");
-
-            grpcStatic.verify(() -> GRPCClient.generateChannel("host", 1234));
-
-            // When closing, ensure the underlying channel is shutdown
-            client.close();
-            verify(channel, times(1)).shutdown();
-            verify(channel, times(1)).awaitTermination(5, TimeUnit.SECONDS);
-        }
-    }
-}
